+++ src/SchedTask.h
@@ -301,6 +301,7 @@ Write your functions:
 
 		9/23/2014 7:44PM initial coding
 		10/01/2020 14:39 revisions for release 1.1.0
+		02/01/2021 14:19 moved constructor and destructor definitions to new file SchedTask.cpp to allow modules to use SchedTask
 */
 
 #ifndef SchedTask_h
@@ -327,10 +328,5 @@ class SchedTask : public SchedBase {
 		virtual void callFunc() {func();}													// call the task on behalf of dispatcher
 		virtual bool checkFunc() {return func != NULL;}									// whether func contains non-NULL
 };
-// Constructor definitions
-SchedTask::SchedTask (unsigned long nxt, unsigned long intval, pFunc fnc) : SchedBase(nxt, intval), func(fnc) {} // constructor definition
-SchedTask::SchedTask (unsigned long nxt, unsigned long intval, long iters, pFunc fnc) : SchedBase(nxt, intval, iters), func(fnc) {} // constructor definition
-SchedTask::SchedTask () : SchedBase(), func(NULL) {} 									// default constructor
-SchedTask::~SchedTask() {;}																	// destructor
 
 #endif